requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
pyyaml>=6.0
numpy>=1.23.0
pytest>=7.0.0
